Added Python 3 support
All that was needed was for the special casing for chunked files to
be dropped, as we do not handle them in any special way (except for
combining the chunks).  Because of the way that this handled
combining bytes, it threw a syntax error in Python 3 and could not
be used.

Installing the package worked fine though for some reason, even
though the setup was throwing the same syntax error.