Add support for CDN_HOST setting
This allows for people to specify the hostname to use for the blob
storage, which is great if you are using a CDN or something else instead
of using Azure's blob storage urls directly.

This also changes how we get settings to return `None` if they aren't
specified, instead of triggering a `KeyError` like we previously would.