Switch to only requiring azure-storage
This switches the requirement to match azure@1.0.3. We no longer
require all of the Azure packages in order to use the storage
adapter. This should not be a breaking change, it only reduces the
scope of the requirements to be as minimal as possible for this
package. There are no functionality changes.